get-rules -> rules-provider
