+++ test/rules-parser.js
@@ -1,5 +1,5 @@
 const assert = require('chai').assert;
-const getRules = require('../src/get-rules.js');
+const RulesProvider = require('../src/rules-provider.js');
 const RulesParser = require('../src/rules-parser.js');
 
 const hasRules = function(rulesOrErrors, expectedRules) {
@@ -29,7 +29,7 @@ const hasErrors = function(rulesOrErrors, expectedErrors) {
 describe('Rule Parser', function() {
   context('When there are rules enabled and disabled', function() {
     it('return the list of enabled rules', function() {
-      const rulesOrErrors = new RulesParser(getRules()).parse({
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide()).parse({
         'no-files-without-scenarios': 'on',
         'no-multiple-empty-lines': 'off',
         'no-trailing-spaces': 'on',
@@ -46,7 +46,7 @@ describe('Rule Parser', function() {
 
   context('when the rule has enabled array configuration', function() {
     it('the rule has the configuration defined in second array item', function() {
-      const rulesOrErrors = new RulesParser(getRules()).parse({
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide()).parse({
         'new-line-at-eof': ['on', 'yes'],
       });
       hasRules(rulesOrErrors, [{
@@ -58,7 +58,7 @@ describe('Rule Parser', function() {
 
   context('when the rule has disabled array configuration', function() {
     it('the rule is not returned', function() {
-      const rulesOrErrors = new RulesParser(getRules()).parse({
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide()).parse({
         'indentation': ['off', {
           'Feature': 1,
           'Background': 1,
@@ -74,7 +74,7 @@ describe('Rule Parser', function() {
 
   describe('Verification fails when', function() {
     it('a non existing rule', function() {
-      const rulesOrErrors = new RulesParser(getRules()).parse({'fake-rule': 'on'});
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide()).parse({'fake-rule': 'on'});
       hasErrors(rulesOrErrors, [{
         type: 'undefined-rule',
         message: 'Rule "fake-rule" does not exist',
@@ -82,7 +82,7 @@ describe('Rule Parser', function() {
     });
 
     it('first item is not "on" or "off"', function() {
-      const rulesOrErrors = new RulesParser(getRules())
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide())
         .parse({'indentation': ['yes', {}]});
       hasErrors(rulesOrErrors, [{
         type: 'config-rule-error',
@@ -91,7 +91,7 @@ describe('Rule Parser', function() {
     });
 
     it('array config does not have 2 items', function() {
-      const rulesOrErrors = new RulesParser(getRules())
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide())
         .parse({'indentation': ['on', {}, 2]});
       hasErrors(rulesOrErrors, [{
         type: 'config-rule-error',
@@ -100,7 +100,7 @@ describe('Rule Parser', function() {
     });
 
     it('a non existing rule sub-config', function() {
-      const rulesOrErrors = new RulesParser(getRules()).parse({
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide()).parse({
         'indentation': ['on', {'featur': 0}],
         'new-line-at-eof': ['on', 'y'],
       });
@@ -116,7 +116,7 @@ describe('Rule Parser', function() {
     });
 
     it('string config is not "on" or "off"', function() {
-      const rulesOrErrors = new RulesParser(getRules())
+      const rulesOrErrors = new RulesParser(new RulesProvider().provide())
         .parse({'indentation': 'no'});
       hasErrors(rulesOrErrors, [{
         type: 'config-rule-error',
