+++ test/rulesdir/rulesdir.js
@@ -2,7 +2,7 @@ const path = require('path');
 const expect = require('chai').expect;
 const Linter = require('../../src/linter/');
 const ConfigProvider = require('../../src/config-provider');
-const getRules = require('../../src/get-rules');
+const RulesProvider = require('../../src/rules-provider');
 const FeaturesProvider = require('../../src/features-provider');
 const RulesParser = require('../../src/rules-parser');
 const NoConfigurableLinter = require('../../src/linter/no-configurable-linter');
@@ -11,7 +11,8 @@ const Gherkin = require('gherkin');
 
 const createLinter = (featureFiles, configPath, additionalRulesDirs) => {
   const configProvider = new ConfigProvider(configPath);
-  const rulesParser = new RulesParser(getRules(additionalRulesDirs));
+  const rulesProvider = new RulesProvider(additionalRulesDirs);
+  const rulesParser = new RulesParser(rulesProvider.provide());
   const featureFinder = new FeaturesProvider(featureFiles);
   const parser = new Gherkin.Parser();
   const noConfigurableFileLinter = new NoConfigurableLinter(parser);
